requirements.txt
basedosdados==1.6.11
geopandas==0.13.2
numpy==1.24.3
pandas==1.4.2
protobuf==3.20.*
requests==2.27.1
pyarrow==6.0.0
python_dateutil==2.9.0.post0
selenium==4.15.2
SQLAlchemy==2.0.17
webdriver_manager==4.0.2
pytz==2023.3
openpyxl==3.0.10
Fiona==1.8.21
Rtree==1.3.0
pygeos==0.14
psycopg2